Move WebsocketConnection wrapper class into KittenWebSocketProvider.cs
